+++ src/main/java/calculator/model/validator/CalculatorValidator.java
@@ -1,4 +1,4 @@
-package calculator.util.validator;
+package calculator.model.validator;
 
 import java.util.Arrays;
 import java.util.concurrent.atomic.AtomicInteger;
