+++ src/test/java/calculator/model/validator/CalculatorValidatorTest.java
@@ -1,5 +1,6 @@
-package calculator.util.validator;
+package calculator.model.validator;
 
+import calculator.model.validator.CalculatorValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
